# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Volaryn API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/activity": {
      "get": {
        "operationId": "activity",
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/admission": {
      "get": {
        "operationId": "admission",
        "parameters": [
          {
            "name": "mint",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admission"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Activity": {
        "type": "object",
        "required": [
          "id",
          "signature",
          "owner",
          "agreement",
          "operation",
          "lastValidBlockHeight",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "agreement": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "format": "int64"
          },
          "createdTerms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Terms"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "lastValidBlockHeight": {
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/Operation"
          },
          "owner": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ActivityPage": {
        "type": "object",
        "required": [
          "items",
          "pending"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          },
          "next": {
            "type": [
              "string",
              "null"
            ]
          },
          "pending": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            },
            "description": "Independent of pagination, so older unresolved operations remain recoverable."
          }
        }
      },
      "Admission": {
        "type": "object",
        "required": [
          "mint",
          "newCommitments",
          "reason"
        ],
        "properties": {
          "mint": {
            "type": "string"
          },
          "newCommitments": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "Operation": {
        "type": "string",
        "enum": [
          "create",
          "activate",
          "exercise",
          "cancel",
          "reclaim",
          "cleanup"
        ]
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "pending",
          "provisional",
          "finalized",
          "failed",
          "expired",
          "reconciled",
          "unresolved"
        ]
      },
      "Terms": {
        "type": "object",
        "required": [
          "underlyingMint",
          "nonce",
          "quantityRaw",
          "payout",
          "premium",
          "acceptBefore",
          "expiresAt"
        ],
        "properties": {
          "acceptBefore": {
            "type": "string"
          },
          "designatedHolder": {
            "type": [
              "string",
              "null"
            ]
          },
          "expiresAt": {
            "type": "string"
          },
          "nonce": {
            "type": "string"
          },
          "payout": {
            "type": "string"
          },
          "premium": {
            "type": "string"
          },
          "quantityRaw": {
            "type": "string"
          },
          "underlyingMint": {
            "type": "string"
          }
        }
      },
...
    }
  }
}